Decompose AI grounding adapter seams
